requirements.txt
streamlit
streamlit-lottie
google-generativeai
requests
pandas
wordcloud
matplotlib
